requirements.txt
# =============================================================================
# IMS CORE - PYTHON DEPENDENCIES (COMPATIBLE)
# =============================================================================

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
pydantic==2.5.0
aio-pika==9.4.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# Caching
redis==5.0.1
hiredis==2.3.2

# API Security
slowapi==0.1.9
python-dotenv==1.0.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Data Validation
pydantic-settings==2.1.0
python-multipart==0.0.6

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Development Tools
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# Security (Linting only - removed safety to avoid conflicts)
bandit==1.7.6

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3
mkdocs-mermaid2-plugin==1.1.1

# Utilities
python-json-logger==2.0.7
rich==13.7.0
typer==0.9.0
psycopg2-binary
# Action Gateway Dependencies
google-generativeai
openai
anthropic
